refactor(A-CodeQuality): Extract cli logic from duke class to CliLauncher class
Closes #29
